Move `iter_symbols` to `symbol` module.
It's now used in multiple places, so `symbol` seems like a more natural
place for it to reside.
